+++ apps/backend/src/modules/contributors/contributor.service.ts
@@ -61,9 +61,14 @@ class ContributorService implements Service {
 	}
 
 	public async findAll(
-		parameters: PaginationQueryParameters,
+		parameters: {
+			contributorName?: string;
+		} & PaginationQueryParameters,
+		hasHidden?: boolean,
 	): Promise<ContributorGetAllResponseDto> {
 		const contributors = await this.contributorRepository.findAll({
+			contributorName: parameters.contributorName ?? "",
+			hasHidden: hasHidden ?? true,
 			page: parameters.page - PAGE_INDEX_OFFSET,
 			pageSize: parameters.pageSize,
 		});
@@ -84,11 +89,23 @@ class ContributorService implements Service {
 		};
 	}
 
-	public async findAllByProjects(
-		projectIds: number[] | undefined,
-	): Promise<ContributorGetAllResponseDto> {
-		const contributors =
-			await this.contributorRepository.findAllByProjects(projectIds);
+	public async findAllByProjectId({
+		contributorName,
+		hasHidden = true,
+		permissionedProjectIds,
+		projectId,
+	}: {
+		contributorName?: string;
+		hasHidden?: boolean;
+		permissionedProjectIds?: number[] | undefined;
+		projectId: number;
+	}): Promise<ContributorGetAllResponseDto> {
+		const contributors = await this.contributorRepository.findAllByProjectId({
+			contributorName: contributorName ?? "",
+			hasHidden,
+			permissionedProjectIds,
+			projectId,
+		});
 
 		return {
 			items: contributors.items.map((item) => {
@@ -106,9 +123,21 @@ class ContributorService implements Service {
 		};
 	}
 
-	public async findAllWithoutPagination(): Promise<ContributorGetAllResponseDto> {
+	public async findAllWithoutPagination({
+		contributorName,
+		hasHidden = true,
+		permissionedProjectIds,
+	}: {
+		contributorName?: string;
+		hasHidden?: boolean;
+		permissionedProjectIds: number[] | undefined;
+	}): Promise<ContributorGetAllResponseDto> {
 		const contributors =
-			await this.contributorRepository.findAllWithoutPagination();
+			await this.contributorRepository.findAllWithoutPagination({
+				contributorName: contributorName ?? "",
+				hasHidden,
+				permissionedProjectIds,
+			});
 
 		return {
 			items: contributors.items.map((item) => {