+++ apps/backend/src/modules/projects/project.repository.ts
@@ -45,7 +45,19 @@ class ProjectRepository implements Repository {
 	}
 
 	public async find(id: number): Promise<null | ProjectEntity> {
-		const item = await this.projectModel.query().findById(id);
+		const item = await this.projectModel
+			.query()
+			.select(
+				"projects.*",
+				"activity_logs.date as analyticsLastSyncedAt",
+				"users.name as analyticsLastSyncedByUser",
+			)
+			.from("projects")
+			.leftJoin("activity_logs", "projects.id", "activity_logs.projectId")
+			.leftJoin("users", "activity_logs.createdByUserId", "users.id")
+			.where("projects.id", id)
+			.orderBy("activity_logs.date", SortType.DESCENDING)
+			.first();
 
 		return item ? ProjectEntity.initialize(item) : null;
 	}