+++ apps/frontend/src/libs/components/icon/libs/maps/icon-name-to-svg.map.ts
@@ -3,6 +3,7 @@ import { type FC } from "react";
 import Access from "~/assets/images/icons/access.svg?react";
 import Analytics from "~/assets/images/icons/analytics.svg?react";
 import Calendar from "~/assets/images/icons/calendar.svg?react";
+import Check from "~/assets/images/icons/check.svg?react";
 import Clipboard from "~/assets/images/icons/clipboard.svg?react";
 import Contributors from "~/assets/images/icons/contributors.svg?react";
 import Cross from "~/assets/images/icons/cross.svg?react";
@@ -19,6 +20,7 @@ import RightDoubleArrow from "~/assets/images/icons/right-double-arrow.svg?react
 import Search from "~/assets/images/icons/search.svg?react";
 import Split from "~/assets/images/icons/split.svg?react";
 import StrikedEye from "~/assets/images/icons/striked-eye.svg?react";
+import Sync from "~/assets/images/icons/sync.svg?react";
 import TrashBin from "~/assets/images/icons/trash-bin.svg?react";
 import Warning from "~/assets/images/icons/warning.svg?react";
 import { type IconName } from "~/libs/types/types.js";
@@ -27,6 +29,7 @@ const iconNameToSvg: Record<IconName, FC<React.SVGProps<SVGSVGElement>>> = {
 	access: Access,
 	analytics: Analytics,
 	calendar: Calendar,
+	check: Check,
 	clipboard: Clipboard,
 	contributors: Contributors,
 	cross: Cross,
@@ -43,6 +46,7 @@ const iconNameToSvg: Record<IconName, FC<React.SVGProps<SVGSVGElement>>> = {
 	search: Search,
 	split: Split,
 	strikedEye: StrikedEye,
+	sync: Sync,
 	trashBin: TrashBin,
 	warning: Warning,
 };