+++ apps/frontend/src/pages/analytics/analytics.tsx
@@ -1,6 +1,11 @@
 import { DateInput, PageLayout, Select } from "~/libs/components/components.js";
-import { DataStatus } from "~/libs/enums/enums.js";
-import { subtractDays } from "~/libs/helpers/helpers.js";
+import { DataStatus, QueryParameterName } from "~/libs/enums/enums.js";
+import {
+	formatDate,
+	getEndOfDay,
+	getStartOfDay,
+	subtractDays,
+} from "~/libs/helpers/helpers.js";
 import {
 	useAppDispatch,
 	useAppForm,
@@ -8,12 +13,18 @@ import {
 	useCallback,
 	useEffect,
 	useFormWatch,
+	useSearchFilters,
+	useSearchParams,
 } from "~/libs/hooks/hooks.js";
 import { actions as activityLogActions } from "~/modules/activity/activity.js";
 
-import { AnalyticsTable } from "./libs/components/components.js";
+import {
+	AnalyticsContributorsSearch,
+	AnalyticsTable,
+} from "./libs/components/components.js";
 import {
 	ANALYTICS_DATE_MAX_RANGE,
+	ANALYTICS_DEFAULT_DATE_RANGE,
 	ANALYTICS_LOOKBACK_DAYS_COUNT,
 } from "./libs/constants/constants.js";
 import { getProjectOptions } from "./libs/helpers/helpers.js";
@@ -27,6 +38,8 @@ const Analytics = (): JSX.Element => {
 		ANALYTICS_LOOKBACK_DAYS_COUNT,
 	);
 
+	const { onSearch, search } = useSearchFilters();
+
 	const { activityLogs, dataStatus, projects } = useAppSelector(
 		({ activityLogs }) => activityLogs,
 	);
@@ -35,33 +48,58 @@ const Analytics = (): JSX.Element => {
 		void dispatch(activityLogActions.loadAllProjects());
 	}, [dispatch]);
 
-	const { control, handleSubmit, isDirty } = useAppForm({
+	const [searchParameters] = useSearchParams();
+	const projectIdQueryParameter = searchParameters.get(
+		QueryParameterName.PROJECT_ID,
+	);
+
+	const { onSearch: onSelect } = useSearchFilters({
+		queryParameterName: QueryParameterName.PROJECT_ID,
+	});
+
+	const { control, errors, handleSubmit } = useAppForm({
 		defaultValues: {
 			dateRange: [
-				subtractDays(todayDate, ANALYTICS_DATE_MAX_RANGE),
+				subtractDays(todayDate, ANALYTICS_DEFAULT_DATE_RANGE),
 				todayDate,
 			] as [Date, Date],
-			project: null,
+			project: projectIdQueryParameter ? Number(projectIdQueryParameter) : null,
+			search,
 		},
 	});
+
+	const handleSearchChange = useCallback(
+		(value: string) => {
+			onSearch(value);
+		},
+		[onSearch],
+	);
 
 	const dateRangeValue = useFormWatch({ control, name: "dateRange" });
 	const projectValue = useFormWatch({ control, name: "project" });
 
+	useEffect(() => {
+		onSelect(projectValue ? String(projectValue) : "");
+	}, [onSelect, projectValue]);
+
 	const handleLoadLogs = useCallback(
-		([startDate, endDate]: [Date, Date], projectId?: null | string) => {
-			const formattedStartDate = startDate.toISOString();
-			const formattedEndDate = endDate.toISOString();
+		([startDate, endDate]: [Date, Date], projectId?: null | number) => {
+			const formattedStartDate = formatDate(
+				getStartOfDay(startDate),
+				"yyyy-MM-dd",
+			);
+			const formattedEndDate = formatDate(getEndOfDay(endDate), "yyyy-MM-dd");
 
 			void dispatch(
 				activityLogActions.loadAll({
+					contributorName: search,
 					endDate: formattedEndDate,
-					projectId: projectId ?? undefined,
+					projectId: projectId?.toString() ?? undefined,
 					startDate: formattedStartDate,
 				}),
 			);
 		},
-		[dispatch],
+		[dispatch, search],
 	);
 
 	useEffect(() => {
@@ -79,12 +117,6 @@ const Analytics = (): JSX.Element => {
 
 	const projectOptions = getProjectOptions(projects);
 
-	useEffect(() => {
-		if (isDirty) {
-			handleFormSubmit();
-		}
-	}, [dateRangeValue, projectValue, isDirty, handleFormSubmit]);
-
 	const isLoading =
 		dataStatus === DataStatus.IDLE || dataStatus === DataStatus.PENDING;
 
@@ -93,6 +125,12 @@ const Analytics = (): JSX.Element => {
 			<h1 className={styles["title"]}>Analytics</h1>
 			<section>
 				<form className={styles["filters-form"]} onSubmit={handleFormSubmit}>
+					<AnalyticsContributorsSearch
+						control={control}
+						errors={errors}
+						name="search"
+						onChange={handleSearchChange}
+					/>
 					<div className={styles["select-wrapper"]}>
 						<Select
 							control={control}
@@ -105,13 +143,15 @@ const Analytics = (): JSX.Element => {
 							placeholder="Select project"
 						/>
 					</div>
-					<DateInput
-						control={control}
-						maxDate={todayDate}
-						maxRange={ANALYTICS_DATE_MAX_RANGE}
-						minDate={minChoosableDate}
-						name="dateRange"
-					/>
+					<div className={styles["date-input-wrapper"]}>
+						<DateInput
+							control={control}
+							maxDate={todayDate}
+							maxRange={ANALYTICS_DATE_MAX_RANGE}
+							minDate={minChoosableDate}
+							name="dateRange"
+						/>
+					</div>
 				</form>
 				<AnalyticsTable
 					activityLogs={activityLogs}