+++ packages/shared/src/libs/helpers/configure-query-string/configure-query-string.helper.ts
@@ -1,0 +1,14 @@
+const configureQueryString = (
+	baseUrl: string,
+	parameters?: ConstructorParameters<typeof URLSearchParams>[number],
+): string => {
+	if (!parameters) {
+		return baseUrl;
+	}
+
+	const searchParameters = new URLSearchParams(parameters);
+
+	return `${baseUrl}?${searchParameters.toString()}`;
+};
+
+export { configureQueryString };