+++ packages/shared/src/libs/helpers/helpers.ts
@@ -1,12 +1,15 @@
 export { addDays } from "./add-days/add-days.helper.js";
 export { changeCase } from "./change-case/change-case.helper.js";
 export { checkHasPermission } from "./check-has-permission/check-has-permission.helper.js";
+export { configureQueryString } from "./configure-query-string/configure-query-string.helper.js";
 export { configureString } from "./configure-string/configure-string.helper.js";
 export { formatDate } from "./format-date/format-date.helper.js";
 export { formatRelativeTime } from "./format-relative-time/format-relative-time.helper.js";
 export { getDateRange } from "./get-date-range/get-date-range.helper.js";
 export { getDifferenceInDays } from "./get-difference-in-days/get-difference-in-days.helper.js";
+export { getEndOfDay } from "./get-end-of-day/get-end-of-day.helper.js";
 export { getRelativeDate } from "./get-relative-date/get-relative-date.helper.js";
 export { getStartOfDay } from "./get-start-of-day/get-start-of-day.helper.js";
+export { getSyncTime } from "./get-sync-time/get-sync-time.helper.js";
 export { initDebounce } from "./init-debounce/init-debounce.helper.js";
 export { subtractDays } from "./subtract-days/subtract-days.helper.js";