+++ packages/shared/src/modules/activity-logs/libs/types/activity-log-get-all-item-analytics-response-dto.type.ts
@@ -1,7 +1,10 @@
 type ActivityLogGetAllItemAnalyticsResponseDto = {
 	commitsNumber: number[];
-	contributorId: string;
-	contributorName: string;
+	contributor: {
+		hiddenAt: null | string;
+		id: string;
+		name: string;
+	};
 };
 
 export { type ActivityLogGetAllItemAnalyticsResponseDto };