+++ packages/shared/src/modules/contributors/libs/types/contributor-patch-response-dto.type.ts
@@ -1,4 +1,5 @@
 type ContributorPatchResponseDto = {
+	hiddenAt: null | string;
 	id: number;
 	name: string;
 };