+++ packages/shared/src/modules/projects/libs/types/project-get-by-id-response-dto.type.ts
@@ -1,4 +1,6 @@
 type ProjectGetByIdResponseDto = {
+	analyticsLastSyncedAt: null | string;
+	analyticsLastSyncedByUser: null | string;
 	apiKey: null | string;
 	description: string;
 	id: number;